refactor: components section별로 폴더 분리
