+++ src/basic/components/manual/Manual.js
