+++ src/basic/components/order-summary/DiscountInfo.js
