+++ src/basic/components/product-selector/StockStatus.js
